% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block}
\alias{block}
\title{Set Multiline Option On or Off}
\usage{
block(value = NULL)
}
\arguments{
\item{value}{Logical. Should comments span over more than one line?}
}
\description{
Toggle the multiline behavior or set it explicitly to TRUE or FALSE.
}
