% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtr_setup.R
\name{cmtr_setup}
\alias{cmtr_setup}
\title{Set cmtr Comment Character(s)}
\usage{
cmtr_setup(chars = c("/*", "*/"), block = TRUE)
}
\arguments{
\item{chars}{Character vector of length 1 or 2, indicating what symbol or
combination of symbols to use as opening, and optionaly closing comment
delimiter.}

\item{block}{Logical. Indicates whether to apply or remove the
comment delimiters on (multiline) blocks (\code{TRUE}), or on a
line-by-line basis (\code{FALSE}).}
}
\description{
Indicate which symbols to use as comment delimiters.
}
